+++ Assets/_Application/Layers/LayerTypes/HierarchicalObject/Properties/TransformPropertySection.cs
@@ -2,6 +2,7 @@ using System;
 using System.Globalization;
 using Netherlands3D.Coordinates;
 using Netherlands3D.Twin.Layers.Properties;
+using Netherlands3D.Twin.UI;
 using RuntimeHandle;
 using UnityEngine;
