+++ Assets/_Application/_Twin/UI/FormField.cs
@@ -1,11 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     [RequireComponent(typeof(TMP_InputField))]
     public class FormField : MonoBehaviour
