+++ Assets/_Application/_Twin/UI/Loader/UI_ProgressIndicator.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Netherlands3D.Twin.UI
+namespace Netherlands3D.Twin.UI.Loader
 {
     public class UI_ProgressIndicator : MonoBehaviour
     {
