+++ Assets/_Application/_Twin/UI/RotatableKnob.cs
@@ -1,11 +1,8 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.EventSystems;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class RotatableKnob : MonoBehaviour, IDragHandler, IBeginDragHandler
     {
