+++ Assets/_Application/_Twin/UI/Snackbar.cs
@@ -1,10 +1,9 @@
 using System.Collections;
 using TMPro;
 using UnityEngine;
-using UnityEngine.Events;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class Snackbar : MonoBehaviour
     {
