+++ Assets/_Functionalities/ObjectInformation/Scripts/BagInspector.cs
@@ -16,7 +16,7 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.Networking;
-using KeyValuePair = Netherlands3D.Twin.KeyValuePair;
+using KeyValuePair = Netherlands3D.Twin.UI.KeyValuePair;
 
 namespace Netherlands3D.Functionalities.ObjectInformation
 {
